New SFO features and UI fixes
- Show thumbnails in the offline view.
- SFO Contextual menu.
- Single and multiple remove from offline.
- Allow open offline files.

+++ MegaApp/common/Services/FolderService.cs
@@ -61,7 +61,8 @@ namespace MegaApp.Services
 
         public static void DeleteFolder (string path)
         {
-            Directory.Delete(path);
+            if(Directory.Exists(path))
+                Directory.Delete(path);
         }
 
         public static void DeleteFolder(string path, bool recursive)
